Add LocalFile implementation for Bookmark, Category repositories
This commit adds the implementation of Bookmark and Category repositories using a local file system. This includes integration tests and updates to controllers, services, and entities. The changes also involve the addition of relevant Lombok dependencies in the build script for test scope. Changes to .gitignore have been made to avoid committing local files used for these implementations. This works under the "local-file" database type property.